+++ app/src/main/java/de/baumann/browser/helper/helper_main.java
@@ -237,6 +237,7 @@ public class helper_main {
         }
 
         sharedPref.edit().putString("openURL", sharedPref.getString("startURL", "https://github.com/scoute-dich/browser/")).apply();
+        sharedPref.edit().putString("webView_url", "").apply();
         sharedPref.edit().putInt("keyboard", 0).apply();
         sharedPref.edit().putInt("tab", 0).apply();
         sharedPref.edit().putInt("closeApp", 0).apply();