+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - fix: overwriting of files while renaming
 - fix: several code improvements
 - improved: layout close button
-- added: override Google AMP-links
+- added: Google encrypted search (override AMP-links)
 
 ### v 4.0
 - removed: donation link on Github